+++ src/baas/cdn/index.js
@@ -1,207 +1,235 @@
 const { Capi } = require('@tencent-sdk/capi')
-const { waitResponse } = require('@ygkit/request')
-const { AddCdnHost, SetHttpsInfo, UpdateCdnConfig, OfflineHost, DeleteCdnHost } = require('./apis')
-const { TIMEOUT, formatCache, formatRefer, getCdnByHost, getPathContent } = require('./utils')
+const { sleep, waitResponse } = require('@ygkit/request')
+const {
+  AddCdnDomain,
+  UpdateDomainConfig,
+  StopCdnDomain,
+  DeleteCdnDomain,
+  PurgeUrlsCache
+} = require('./apis')
+const {
+  TIMEOUT,
+  formatCertInfo,
+  formatOrigin,
+  camelCaseProperty,
+  getCdnByDomain,
+  flushEmptyValue
+} = require('./utils')
 
 class Cdn {
   constructor(credentials = {}) {
     this.credentials = credentials
-  }
 
-  async deploy(inputs = {}) {
-    const {
-      host,
-      hostType,
-      origin,
-      oldState = {},
-      backupOrigin = '',
-      serviceType = 'web',
-      fullUrl = 'off',
-      fwdHost,
-      cache,
-      cacheMode = 'simple',
-      refer,
-      accessIp,
-      https
-    } = inputs
-
-    const capi = new Capi({
+    this.capi = new Capi({
       AppId: this.credentials.AppId,
       SecretId: this.credentials.SecretId,
       SecretKey: this.credentials.SecretKey,
       Token: this.credentials.Token
     })
+  }
 
-    const cdnInputs = {
-      host: host,
-      projectId: 0,
-      hostType: hostType,
-      origin: origin,
-      backupOrigin: backupOrigin,
-      serviceType: serviceType,
-      fullUrl: fullUrl,
-      fwdHost: fwdHost || host,
-      cacheMode: cacheMode
-    }
-
-    if (cache) {
-      cdnInputs.cache = JSON.stringify(formatCache(cache))
-    }
-    if (refer) {
-      cdnInputs.refer = JSON.stringify(formatRefer(refer[0]))
-    }
-    if (accessIp) {
-      cdnInputs.accessIp = JSON.stringify(accessIp)
-    }
-
-    const cdnInfo = await getCdnByHost(capi, host)
-
-    const sourceInputs = JSON.parse(JSON.stringify(inputs))
-    const sourceHttpsConf = sourceInputs.https
-    const sourceBaseConf = {}
-    Object.keys(sourceInputs).forEach((key) => {
-      if (key !== 'https') {
-        sourceBaseConf[key] = sourceInputs[key]
-      }
+  async purgeCdnUrls(urls) {
+    console.log(`Purging CDN caches, it will work in 5 minutes...`)
+    await PurgeUrlsCache(this.capi, {
+      Urls: urls
     })
+  }
+
+  async offlineCdnDomain(domain) {
+    const { Status } = await getCdnByDomain(this.capi, domain)
+    if (Status === 'online') {
+      // disable first
+      await StopCdnDomain(this.capi, { Domain: domain })
+    } else if (Status === 'processing') {
+      throw new Error(`Status is not operational for ${domain}`)
+    }
+  }
+
+  async deploy(inputs = {}) {
+    const { oldState = {} } = inputs
+    delete inputs.oldState
+    const {
+      Async = false,
+      Domain,
+      Origin,
+      ServiceType = 'web',
+      Area = 'mainland',
+      Https,
+      Cache,
+      IpFilter,
+      IpFreqLimit,
+      StatusCodeCache,
+      ForceRedirect,
+      Compression,
+      BandwidthAlert,
+      RangeOriginPull,
+      FollowRedirect,
+      ErrorPage,
+      RequestHeader,
+      ResponseHeader,
+      DownstreamCapping,
+      CacheKey,
+      ResponseHeaderCache,
+      VideoSeek,
+      OriginPullOptimization,
+      Authentication,
+      Seo,
+      Referer,
+      MaxAge,
+      SpecificConfig,
+      OriginPullTimeout,
+      RefreshCdn
+    } = camelCaseProperty(inputs)
+
+    const cdnInputs = flushEmptyValue({
+      Domain,
+      Origin: formatOrigin(Origin),
+      ServiceType,
+      Area,
+      Cache,
+      IpFilter,
+      IpFreqLimit,
+      StatusCodeCache,
+      ForceRedirect,
+      Compression,
+      BandwidthAlert,
+      RangeOriginPull,
+      FollowRedirect,
+      ErrorPage,
+      RequestHeader,
+      ResponseHeader,
+      DownstreamCapping,
+      CacheKey,
+      ResponseHeaderCache,
+      VideoSeek,
+      OriginPullOptimization,
+      Authentication,
+      Seo,
+      Referer,
+      MaxAge,
+      SpecificConfig,
+      OriginPullTimeout
+    })
+
     const outputs = {
-      host: host,
-      origin: origin,
-      cname: `${host}.cdn.dnsv1.com`,
-      inputCache: {
-        base: JSON.stringify(sourceBaseConf),
-        https: JSON.stringify(sourceHttpsConf)
+      https: false,
+      domain: Domain,
+      origins: cdnInputs.Origin.Origins,
+      cname: `${Domain}.cdn.dnsv1.com`,
+      inputCache: JSON.stringify(cdnInputs)
+    }
+
+    if (Https) {
+      outputs.https = true
+      cdnInputs.Https = {
+        Switch: Https.Switch,
+        Http2: Https.Http2 || 'off',
+        OcspStapling: Https.OcspStapling || 'off',
+        VerifyClient: Https.VerifyClient || 'off',
+        CertInfo: formatCertInfo(Https.CertInfo)
       }
     }
+    if (ForceRedirect && Https) {
+      cdnInputs.ForceRedirect = {
+        Switch: ForceRedirect.Switch,
+        RedirectStatusCode: ForceRedirect.RedirectStatusCode || 301,
+        RedirectType: 'https'
+      }
+    }
+
+    const cdnInfo = await getCdnByDomain(this.capi, Domain)
+
+    const sourceInputs = JSON.parse(JSON.stringify(cdnInputs))
 
     const createOrUpdateCdn = async () => {
       if (cdnInfo) {
         // update
-        console.log(`The CDN domain ${host} has existed.`)
+        console.log(`The CDN domain ${Domain} has existed.`)
         console.log('Updating...')
-        cdnInputs.hostId = cdnInfo.id
-        await UpdateCdnConfig(capi, cdnInputs)
-        outputs.hostId = cdnInfo.id
+        // when update, can not set ServiceType parameter
+        await UpdateDomainConfig(this.capi, cdnInputs)
+        outputs.resourceId = cdnInfo.ResourceId
       } else {
         // create
-        console.log(`Adding CDN domain ${host}...`)
+        console.log(`Adding CDN domain ${Domain}...`)
         try {
-          await AddCdnHost(capi, cdnInputs)
+          await AddCdnDomain(this.capi, cdnInputs)
         } catch (e) {
           if (e.code === 9111) {
             console.log(`Please goto https://console.cloud.tencent.com/cdn open CDN service.`)
           }
           throw e
         }
-        const { id } = await getCdnByHost(capi, host)
+        await sleep(1000)
+        const detail = await getCdnByDomain(this.capi, Domain)
+
         outputs.created = true
-        outputs.hostId = id
+        outputs.resourceId = detail && detail.ResourceId
       }
 
-      // state=4: deploying status, we can not do any operation
-      console.log('Waiting for CDN deploy success...')
-      await waitResponse({
-        callback: async () => getCdnByHost(capi, host),
-        targetProp: 'status',
-        targetResponse: 5,
-        timeout: TIMEOUT
-      })
-      console.log(`CDN deploy success to host: ${host}`)
-    }
+      console.log('Waiting for CDN deploy success, it maybe cost 5 minutes....')
+      // When set syncFlow false, just continue, do not wait for online status.
+      if (Async === false) {
+        await waitResponse({
+          callback: async () => getCdnByDomain(this.capi, Domain),
+          targetProp: 'Status',
+          targetResponse: 'online',
+          timeout: TIMEOUT
+        })
 
-    const creatOrUpdateHttps = async () => {
-      if (https) {
-        console.log(`Setup https for ${host}...`)
-        // update https
-        const httpsInputs = {
-          host: host,
-          httpsType: https.httpsType,
-          forceSwitch: https.forceSwitch,
-          http2: https.http2
+        // refresh cdn urls
+        if (RefreshCdn && RefreshCdn.Urls) {
+          await this.purgeCdnUrls(RefreshCdn.Urls)
         }
-        // if set certId, it is prefered
-        if (https.certId) {
-          httpsInputs.certId = https.certId
-        } else {
-          const certContent = await getPathContent(https.cert)
-          const privateKeyContent = await getPathContent(https.privateKey)
-          httpsInputs.cert = certContent
-          httpsInputs.privateKey = privateKeyContent
-        }
-
-        await SetHttpsInfo(capi, httpsInputs)
-        outputs.https = true
-      } else {
-        console.log(`Removing https for ${host}...`)
-        // delete https
-        const httpsInputs = {
-          host: host,
-          httpsType: 0
-        }
-        await SetHttpsInfo(capi, httpsInputs)
-        outputs.https = false
       }
-      await waitResponse({
-        callback: async () => getCdnByHost(capi, host),
-        targetProp: 'status',
-        targetResponse: 5,
-        timeout: TIMEOUT
-      })
+      console.log(`CDN deploy success to domain: ${Domain}`)
     }
 
     // pass state for cache check
     const { inputCache } = oldState
-    if (inputCache && inputCache.base === JSON.stringify(sourceBaseConf)) {
-      console.log(`No base configuration changes for CDN domain ${host}`)
-      outputs.hostId = cdnInfo.id
+    if (inputCache && inputCache === JSON.stringify(sourceInputs)) {
+      console.log(`No configuration changes for CDN domain ${Domain}`)
+      outputs.resourceId = cdnInfo.ResourceId
     } else {
       await createOrUpdateCdn()
-    }
-
-    if (inputCache && inputCache.https === JSON.stringify(sourceHttpsConf)) {
-      console.log(`No https configuration changes for CDN domain ${host}`)
-      outputs.https = !!sourceHttpsConf
-    } else {
-      await creatOrUpdateHttps()
     }
 
     return outputs
   }
 
   async remove(inputs = {}) {
-    const capi = new Capi({
-      SecretId: this.credentials.SecretId,
-      SecretKey: this.credentials.SecretKey,
-      Token: this.credentials.Token
-    })
-
-    const { host } = inputs
-    if (!host) {
-      throw new Error('host is required')
+    const { domain } = inputs
+    if (!domain) {
+      throw new Error('domain is required')
     }
 
-    // need circle for deleting, after host status is 6, then we can delete it
-    console.log(`Start removing CDN for ${host}`)
-    const { status } = await getCdnByHost(capi, host)
-    // status=5: online
-    // state=4: deploying
-    // state=6: offline
-    if (status === 5) {
+    // need circle for deleting, after domain status is 6, then we can delete it
+    console.log(`Start removing CDN for ${domain}`)
+    const detail = await getCdnByDomain(this.capi, domain)
+    if (!detail) {
+      console.log(`CDN domain ${domain} not exist`)
+      return {}
+    }
+
+    const { Status } = detail
+
+    if (Status === 'online') {
       // disable first
-      await OfflineHost(capi, { host: host })
-      console.log(`Waiting for offline ${host}...`)
-      await waitResponse({
-        callback: async () => getCdnByHost(capi, host),
-        targetProp: 'status',
-        targetResponse: 6,
-        timeout: TIMEOUT
-      })
-    } else if (status === 4) {
-      throw new Error(`Status is not operational for ${host}`)
+      await StopCdnDomain(this.capi, { Domain: domain })
+    } else if (Status === 'processing') {
+      console.log(`Status is not operational for ${domain}`)
+      return {}
     }
-    console.log(`Removing CDN for ${host}`)
-    await DeleteCdnHost(capi, { host: host })
-    console.log(`Removed CDN for ${host}.`)
+    console.log(`Waiting for offline ${domain}...`)
+    await waitResponse({
+      callback: async () => getCdnByDomain(this.capi, domain),
+      targetProp: 'Status',
+      targetResponse: 'offline',
+      timeout: TIMEOUT
+    })
+    console.log(`Removing CDN for ${domain}`)
+    await DeleteCdnDomain(this.capi, { Domain: domain })
+    console.log(`Removed CDN for ${domain}.`)
     return {}
   }
 }