+++ src/baas/scf/index.js
@@ -497,16 +497,33 @@ class Scf {
 
   // 移除函数的主逻辑
   async remove(inputs = {}) {
-    console.log(`Deleteing funtion ${inputs.functionName || inputs.FunctionName} ...`)
-    await this.deleteFunction(
-      inputs.functionName || inputs.FunctionName,
-      inputs.namespace || inputs.Namespace || defaultNamespace
-    )
+    console.log(`Deleteing function ${inputs.functionName || inputs.FunctionName} ...`)
+    const functionName = inputs.functionName || inputs.FunctionName
+    const namespace = inputs.namespace || inputs.Namespace || defaultNamespace
+
+    // check function exist, then delete
+    const func = await this.getFunction(namespace, functionName)
+
+    if (!func) {
+      console.log(`Function ${functionName} not exist...`)
+      return
+    }
+
+    if (func.Status === 'Updating' || func.Status === 'Creating') {
+      console.log(`Function ${functionName} status is ${func.Status}, can not delete...`)
+      return
+    }
+
+    await this.deleteFunction(functionName, namespace)
 
     if (inputs.Triggers) {
       for (let i = 0; i < inputs.Triggers.length; i++) {
         if (inputs.Triggers[i].serviceId) {
-          await this.deleteAPIGW(inputs.Triggers[i])
+          try {
+            await this.deleteAPIGW(inputs.Triggers[i])
+          } catch (e) {
+            console.log(e)
+          }
         }
       }
     }